Use portable printf format specifiers.
Integers were being printed with format specifiers "lu" and "ld",
indicating unsigned and signed "long" values, which are of
indeterminate size.  The code wisely uses precise
sizes (e.g. uint32_t) througout, but when these are printed with the
indeterminate-sized format specifies, this may result in a mismatch
between the formatter's assumed integer size and the actual integer
size.

To fix this, this change replaces the "lu" and "ld" specifiers with
the PRIu32 and PRId32 macros from <cinttypes>.