Use portable max value.
The code uses LONG_MAX to indicate that an integer has an
indeterminate value, but the values in question are specified as
uint32_ts, which may or may not be able to hold the maximum value of a
long, depending on the size of a long on the target platform.

I added a template function "maxOf" that uses std::numeric_limits to
return the maximum possible value for a given variable, and replaced
uses of LONG_MAX with that.